Added Validation for Auth Fields and displayed error Messages
Added validation for sign-up fields (email validation, password format validation, confirm password check), login fields, change password fields and update profile fields.

Error message displays added for incorrect login credentials, and incorrect credentials for password change.

CustomizedTextInput  component to provide for required fields, display error message, and a footer message (as seen in the password field). Important for future.

Disabled submit of all the above forms where information is needed to be filled or where a change is necessary to submit (for e.g. in the profile update, the submit is only enabled if the fullname or lastname is mutated).